+++ Paradigm/test/utopia/paradigm/test/ColorTest.scala
@@ -21,5 +21,9 @@ object ColorTest extends App
 			assert(avg.values.forall { case (_, v) => v >= 0 && v <= 255 }, s"Avg of $inputs is $avg")
 		}
 	
-	println("Success!")
+	val colorSet = Set(Hsl(Angle.degrees(21.06)), Hsl(Angle.degrees(280)))
+	val c = Color.average(Set(Hsl(Angle.degrees(21.06)), Hsl(Angle.degrees(280))))
+	assert(c.values.forall { case (_, v) => v >= 0 && v <= 255 }, s"Avg of $colorSet is $c")
+	
+	println(c)
 }